refactor(launch): split LaunchSubTabs into launch-subtabs/* submodules
